updated naming from BQDS to CDS and readme change
