+++ client/cmd/cdmc/LICENSE
