+++ client/cmd/cdmc/cmd/publish.go
@@ -23,8 +23,8 @@ import (
 // multicastPublishCmd represents the mlcstPublish command
 var multicastPublishCmd = &cobra.Command{
 	Use:   "publish",
-	Short: "BQDS client multicast publisher service",
-	Long: `The BQDS client multicast publisher service will listen to a specific multicast interface and address, then publish those messages to the specified PubSub Topic. For example:
+	Short: "CDS client multicast publisher service",
+	Long: `The CDS client multicast publisher service will listen to a specific multicast interface and address, then publish those messages to the specified PubSub Topic. For example:
 
 -n "udp"
 -a "239.0.0.0:9999"
