+++ client/cmd/cdmc/cmd/socket.go
@@ -28,8 +28,8 @@ var (
 // socketCmd represents the socket command
 var socketCmd = &cobra.Command{
 	Use:   "socket",
-	Short: "BQDS client socket service",
-	Long: `The BQDS client socket service will attach to a raw unix or tcp socket and iterate over
+	Short: "CDS client socket service",
+	Long: `The CDS client socket service will attach to a raw unix or tcp socket and iterate over
 stream data if new line. For example:
 
   -n "unix"
