+++ client/files/images/cds-multicast-client-cloud-architecture.png
